+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ### Bug fixes
 
--   None
+-   Replace PodDisruptionBudget if spec changes (https://github.com/pulumi/pulumi-kubernetes/pull/527)
 
 ## 0.22.0 (March 25, 2019)